+++ src/main/java/me/pajic/misctweaks/MiscTweaks.java
@@ -19,8 +19,6 @@ import me.pajic.misctweaks.platform.fabric.FabricPlatform;
 public class MiscTweaks {
 
 	public static final String MOD_ID = /*$ mod_id*/ "misctweaks";
-	public static final String MOD_VERSION = /*$ mod_version*/ "1.2.9";
-	public static final String MOD_FRIENDLY_NAME = /*$ mod_name*/ "MiscTweaks";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static final Identifier CONFIG_RL = id("config");
 	public static ModConfig CONFIG = ConfigApiJava.registerAndLoadConfig(ModConfig::new);
@@ -43,17 +41,7 @@ public class MiscTweaks {
 	}
 
 	public static Identifier id(String path) {
-		//? if 1.20.1
-		//return new ResourceLocation(MOD_ID, path);
-		//? if > 1.20.1
 		return Identifier.fromNamespaceAndPath(MOD_ID, path);
-	}
-
-	public static Identifier vanillaId(String path) {
-		//? if 1.20.1
-		//return new ResourceLocation(path);
-		//? if > 1.20.1
-		return Identifier.withDefaultNamespace(path);
 	}
 
 	public static void debugLog(String message, Object ... args) {