+++ src/main/java/me/pajic/misctweaks/mixin/PlayerMixin.java
@@ -7,10 +7,9 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import net.minecraft.core.component.DataComponents;
 //? if < 1.21.10
 //import net.minecraft.world.item.ElytraItem;
-//? if > 1.20.1
-import net.minecraft.core.component.DataComponents;
 
 @Mixin(Player.class)
 public class PlayerMixin {
@@ -24,10 +23,7 @@ public class PlayerMixin {
         if (
 				MiscTweaks.CONFIG.elytraSwimTweak.get() &&
                 self.isFallFlying() &&
-                //? if >= 1.21.1
 				self.isInLiquid() &&
-                //? if < 1.21.1
-                //self.isInWaterOrBubble() || self.isInLava() &&
                 //? if <= 1.21.1
                 //self.getItemBySlot(EquipmentSlot.CHEST).getItem() instanceof ElytraItem
                 //? if > 1.21.1