+++ src/main/java/me/pajic/misctweaks/mixin/SweetBerryBushBlockMixin.java
@@ -2,16 +2,19 @@ package me.pajic.misctweaks.mixin;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import me.pajic.misctweaks.MiscTweaks;
+import net.minecraft.util.Mth;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.SweetBerryBushBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-//? if < 1.21.1
-//import net.minecraft.world.item.ArmorItem;
 //? if >= 1.21.1 {
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.ItemTags;
@@ -35,21 +38,33 @@ public class SweetBerryBushBlockMixin {
     //? if > 1.21.1
     private boolean preventDamageIfSneakingOrWearingLegArmor(Entity instance, ServerLevel serverLevel, DamageSource source, float amount, Operation<Boolean> original) {
         if (instance instanceof Player p) {
-            if (
-                    (MiscTweaks.CONFIG.sneakingPreventsBerryBushDamage.get() && p.isShiftKeyDown()) ||
-                    (MiscTweaks.CONFIG.legArmorPreventsBerryBushDamage.get() &&
-							//? if >= 1.21.1
-							p.getItemBySlot(EquipmentSlot.LEGS).is(ItemTags.LEG_ARMOR)
-                            //? if < 1.21.1
-							//p.getItemBySlot(EquipmentSlot.LEGS).getItem() instanceof ArmorItem ai && ai.getType() == ArmorItem.Type.LEGGINGS
-                    )
-            ) {
-                return false;
-            }
+            if ((MiscTweaks.CONFIG.sneakingPreventsBerryBushDamage.get() && p.isShiftKeyDown())) return false;
         }
+		if (instance instanceof LivingEntity l && MiscTweaks.CONFIG.armorPreventsBerryBushDamage.get()) {
+			if (l.getItemBySlot(EquipmentSlot.LEGS).is(ItemTags.LEG_ARMOR) || l.hasItemInSlot(EquipmentSlot.BODY)) return false;
+		}
         //? if <= 1.21.1
         //return original.call(instance, source, amount);
         //? if > 1.21.1
         return original.call(instance, serverLevel, source, amount);
     }
+
+	@WrapOperation(
+			method = "entityInside",
+			at = @At(
+					value = "INVOKE",
+					target = "Lnet/minecraft/world/entity/Entity;makeStuckInBlock(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/phys/Vec3;)V"
+			)
+	)
+	private void reduceSlowEffectWhenWearingLegArmor(Entity instance, BlockState state, Vec3 motionMultiplier, Operation<Void> original) {
+		if (MiscTweaks.CONFIG.armorReducesBerryBushSlow.get() && instance instanceof LivingEntity l) {
+			if (l.getItemBySlot(EquipmentSlot.LEGS).is(ItemTags.LEG_ARMOR) || l.hasItemInSlot(EquipmentSlot.BODY)) {
+				double d = MiscTweaks.CONFIG.berryBushSlowReduction.get();
+				if (d == 2.2) return;
+				original.call(instance, state, motionMultiplier.multiply(d, d, d));
+				return;
+			}
+		}
+		original.call(instance, state, motionMultiplier);
+	}
 }